+++ pronto_ros/src/gps_ros_handler.cpp
@@ -1,4 +1,4 @@
-#include "pronto_estimator_ros/gps_ros_handler.hpp"
+#include "pronto_ros/gps_ros_handler.hpp"
 
 namespace MavStateEst {
